--- config/initializers/selenium.rb
@@ -1,10 +1,0 @@
-# require 'selenium-webdriver'
-
-# Capybara.register_driver :selenium do |app|
-#   options = Selenium::WebDriver::Firefox::Options.new
-#   options.headless = false
-
-#   Capybara::Selenium::Driver.new(app, browser: :firefox, options: options)
-# end
-
-# Capybara.default_driver = :selenium
