+++ spec/rails_helper.rb
@@ -67,10 +67,10 @@ RSpec.configure do |config|
     options.add_argument('--headless')
     options.add_argument('--no-sandbox')
     options.add_argument('--disable-dev-shm-usage')
-  
-    Capybara::Selenium::Driver.new(app, browser: :chrome, options: options)
+
+    Capybara::Selenium::Driver.new(app, browser: :chrome, options:)
   end
-  
+
   Capybara.javascript_driver = :selenium_chrome_headless
   Capybara.server = :puma
-  end
+end
